Move dbaccessor fixture
