+++ sardes/plugins/conftest.py
@@ -7,25 +7,4 @@
 # Licensed under the terms of the GNU General Public License.
 # -----------------------------------------------------------------------------
 
-
-# ---- Standard imports
-import os
-import os.path as osp
-os.environ['SARDES_PYTEST'] = 'True'
-
-# ---- Third party imports
-import pytest
-
-# ---- Local imports
-from sardes.database.accessors import DatabaseAccessorSardesLite
 from sardes.database.accessors.tests.conftest import *
-
-
-@pytest.fixture
-def dbaccessor(tmp_path, database_filler):
-    dbaccessor = DatabaseAccessorSardesLite(
-        osp.join(tmp_path, 'sqlite_database_test.db'))
-    dbaccessor.init_database()
-    database_filler(dbaccessor)
-
-    return dbaccessor
